kernel: Don't update defconfig with KERNEL_ADDITIONAL_CONFIG options
If 'make kernelconfig' is run in an environment where
KERNEL_ADDITIONAL_CONFIG is set, the options in the additional
defconfig will be saved back to the defconfig when savedefconfig
is performed.

While not really a bug, this increases the likelihood of someone
accidentally committing these changes or getting confused by
unexpected deltas added to their defconfigs.

Fix this by skipping the merge-config.sh step for the 'kernelconfig'
make target. As a side-effect, options in KERNEL_ADDITIONAL_CONFIG
will also not be displayed in menuconfig. This is not easily
avoidable, and arguably not wrong anyway.

Change-Id: I920be0338bb8afc87f6061c55aafb27ba658974d
Signed-off-by: D. Andrei Măceș <andrei@unlegacy-android.org>